Move ini config parsing to its own module
This is essentially in preparation for a proof-of-concept
implementation of #2938 I am working on, which will want to
call into this code from build, and a build -> main dependency
seems wrong.
