+++ mypy/test/testfinegrained.py
@@ -32,7 +32,7 @@ from mypy.test.helpers import (
 from mypy.server.mergecheck import check_consistency
 from mypy.dmypy_util import DEFAULT_STATUS_FILE
 from mypy.dmypy_server import Server
-from mypy.main import parse_config_file
+from mypy.config_parser import parse_config_file
 from mypy.find_sources import create_source_list
 
 import pytest  # type: ignore  # no pytest in typeshed
